feat(kernel): implement kernel preemption, namespace system themes, and resolve static analysis warnings
- Implement kernel preemption tracking via per-process preemption counters and pending reschedule flags
- Defer timer interrupt context switches when preemption is disabled and trigger a deferred yield upon enabling preemption
- Update all spinlock primitives to automatically disable and enable local core preemption during lock acquisition and release
- Encapsulate system theme definitions inside a kernel namespace and refactor global color macros to use the scoped theme pointer
- Resolve static analysis warnings by adding cppcheck suppressions and replacing C-style pointer casts with explicit reinterpret_casts
- Prevent potential out-of-bounds writes in display composition by capping the damage rectangle loop against a max local limit
- Fix compiler and print format mismatches by explicitly casting 64-bit sizes, offsets, and timestamps to unsigned long long
- Secure internal application state structures within anonymous namespaces across system utilities
- Avoid a potential null pointer dereference in menubar rendering by validating the configuration registry pointer before use
- Replace a raw memset call in the calculator utility with modern C++ empty brace value initialization
- Unconditionally parse mouse wheel data from HID reports instead of checking report length bounds

+++ src/usr/libc/log.c
@@ -36,7 +36,8 @@ void log_vmessage(LogLevel level, const char *scope, const char *fmt, va_list ap
     uint64_t millis = ticks % 1000;
 
     char line[896];
-    int line_len = snprintf(line, sizeof(line), "[%llu.%03llu] %-4s %s %s", seconds, millis, resolved_scope,
+    int line_len = snprintf(line, sizeof(line), "[%llu.%03llu] %-4s %s %s",
+                            (unsigned long long)seconds, (unsigned long long)millis, resolved_scope,
                             level_mark(level), message);
     if (line_len <= 0)
         return;
